1.0.0.2
Fixed handling of ListUrl parameter, list name was not being extracted correctly, now looks for the element after the /Lists/ segment. Note that this part of the URL needs to be set manually to the name of the list, if different from the path (i.e. a renamed list) it still breaks. Better for vNext to add a ListName parm and correctly translate the path to the List Name or GUID.